Redefininh project structure
