+++ CONTRIBUTING.md
@@ -1,0 +1,2 @@
+# OSS contributors guide
+# ...original content to be added...
