+++ LICENSE
@@ -1,0 +1,2 @@
+# AGPL/SSPL License
+# Placeholder for license text
