+++ README.md
@@ -1,3 +1,17 @@
+# GovernAble: AI Security Platform
+
+## Problem
+AI requires optimal and efficient data to be trained. This implies that the data is free from PII or confidential information that might be used and kept by the LLM or Machine as context to be used.
+
+## Solution
+Technical solution which is lightweight and can be integrated via plugin into multiple models for verification, scanning and warn the files for PII content.
+
+## New Structure
+See repository for new modular structure (engine, api, proxy, web, enterprise, tests).
+
+---
+
+Original content:
 # AI Security PII detection tool
 ## Problem 
 AI requires optimal and efficient data to be trained . This implies that the data is free from PII or confidential information that might be used and kept by the LLM or Machine as context to be used .
