+++ api/config.py
@@ -1,0 +1,2 @@
+# Common config for GovernAble API
+# ...original content to be added...
