--- app/api/v1/secrets.py
@@ -1,1 +1,0 @@
-.
