--- app/core/pii_scanner.py
@@ -1,1 +1,0 @@
-.
