--- app/utils/auth.py
@@ -1,1 +1,0 @@
-.
