+++ engine/rules/base_patterns.yml
@@ -1,0 +1,2 @@
+# Default open-core patterns for GovernAble engine
+# ...original content to be added...
