+++ enterprise/integrations/.keep
@@ -1,0 +1,1 @@
+# Placeholder for integrations
