--- tests/test_file_parser.py
@@ -1,1 +1,0 @@
-.
