+++ tests/test_proxy.ts
@@ -1,0 +1,2 @@
+// Shared test for proxy
+// ...original content to be added...
