--- tests/test_secrets.py
@@ -1,1 +1,0 @@
-.
