+++ web/backend/.keep
@@ -1,0 +1,1 @@
+# Placeholder for backend app
