+++ web/frontend/.keep
@@ -1,0 +1,1 @@
+# Placeholder for frontend app
